refactor: rename userTiming and elementTiming folders to user-timing and element-timing
