+++ packages/web-sdk/src/instrumentations/element-timing/index.ts
